Added aconv ffmpeg task to Grunt build
The task 'aconv' uses the command line ffmpeg tool to convert
existing .ogg to .m4a (AAC) format files.

To convert to AAC format files it is required that ffmpeg is already
compiled with --enable-nonfree and --enable-libfdk-aac. Note that
fdk is the best possible free option. For more info check here:
http://trac.ffmpeg.org/wiki/AACEncodingGuide#WhichencodershouldIuseWhatprovidesthebestquality